Fix crash when printing error messages from ds4drv
From ds4drv, arguments like:

msg = 'Unable to open device: {0}'
args = 'device or resource busy'

are passed. Program crashes due to formatting error when this type of argument
is passed.

As a sidenote, this message was observed when trying to connect to the same
device from two different rosmasters. To see this, run:

$ export ROS_MASTER_URI='http://host1:11311'
$ roslaunch ds4_driver ds4_driver.launch

and in a different terminal,

$ export ROS_MASTER_URI='http://host2:11311'
$ roslaunch ds4_driver ds4_driver.launch